added `.float()` before `torch.std`

+++ GANDLF/data/augmentation/noise_enhanced.py
@@ -66,5 +66,5 @@ class RandomNoiseEnhanced(RandomTransform, IntensityTransform):
         std_ranges = self.std_original
         if self.std_original is None:
             # calculate the default std range based on 1.5% of the input image std - https://github.com/mlcommons/GaNDLF/issues/518
-            std_ranges = (0, 0.015 * torch.std(image.data).item())
+            std_ranges = (0, 0.015 * torch.std(image.data.float()).item())
         return self._parse_range(std_ranges, "std", min_constraint=0)
